# services/api/docs/swagger.yaml
basePath: /
definitions:
  internal_handler.CacheHealthStatus:
    properties:
      status:
        enum:
        - ok
        - unreachable
        example: ok
        type: string
    type: object
  internal_handler.DBHealthStatus:
    properties:
      acquired_conns:
        example: 2
        type: integer
      idle_conns:
        example: 8
        type: integer
      max_conns:
        example: 50
        type: integer
      status:
        enum:
        - ok
        - unreachable
        example: ok
        type: string
      total_conns:
        example: 10
        type: integer
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        example: query timeout
        type: string
      retry_after:
        example: 1
        type: integer
    type: object
  internal_handler.HealthResponse:
    properties:
      cache:
        $ref: '#/definitions/internal_handler.CacheHealthStatus'
      db:
        $ref: '#/definitions/internal_handler.DBHealthStatus'
      status:
        enum:
        - healthy
        - degraded
        - unhealthy
        example: healthy
        type: string
    type: object
  signal_api_internal_repository.PageMeta:
    properties:
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
    type: object
  signal_api_internal_repository.Scenario:
    properties:
      confidence:
        example: 0.85
        maximum: 1
        minimum: 0
        type: number
      direction:
        enum:
        - long
        - short
        example: long
        type: string
      entry_point:
        example: 49500
        type: number
      entry_type:
        enum:
        - market
        - limit
        example: limit
        type: string
      id:
        example: 1
        type: integer
      invalidation:
        example: Daily close below 47000 invalidates setup
        type: string
      reasoning:
        example: Strong support level with increasing volume and RSI divergence
        type: string
      results:
        items:
          $ref: '#/definitions/signal_api_internal_repository.ScenarioResult'
        type: array
      signal_id:
        example: 1
        type: integer
      status:
        enum:
        - active
        - closed
        - invalidated
        example: active
        type: string
      stop_loss:
        example: 47000
        type: number
      take_profits:
        type: object
    type: object
  signal_api_internal_repository.ScenarioResult:
    properties:
      evaluated_at:
        example: "2024-01-16T14:30:00Z"
        type: string
      hit_sl:
        type: number
      hit_tp:
        example: 51000
        type: number
      id:
        example: 1
        type: integer
      max_drawdown:
        example: -2.3
        type: number
      pnl_percent:
        example: 8.5
        type: number
      result:
        enum:
        - win
        - loss
        - partial
        - pending
        example: win
        type: string
      scenario_id:
        example: 1
        type: integer
    type: object
  signal_api_internal_repository.Signal:
    properties:
      ai_summary:
        example: Strong bullish breakout confirmed with high volume
        type: string
      analyzed_at:
        example: "2024-01-15T09:05:00Z"
        type: string
      created_at:
        example: "2024-01-15T09:00:00Z"
        type: string
      current_market_price:
        type: object
      id:
        example: 1
        type: integer
      raw_text:
        example: BTC breaks key resistance — targeting 52k
        type: string
      scenarios:
        items:
          $ref: '#/definitions/signal_api_internal_repository.Scenario'
        type: array
      source:
        type: object
      symbol:
        example: BTCUSDT
        type: string
      trace_id:
        example: 3f2d1a9b-feed-4c2e-bae4-abc123def456
        type: string
    type: object
  signal_api_internal_repository.SignalsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/signal_api_internal_repository.Signal'
        type: array
      meta:
        $ref: '#/definitions/signal_api_internal_repository.PageMeta'
    type: object
host: localhost:8080
info:
  contact:
    name: Signal Team
  description: Read-only REST API for accessing AI-analyzed trading signals, scenario
    plans, and result tracking.
  title: Signal API
  version: "1.0"
paths:
  /health:
    get:
      description: |-
        Returns the overall service health including PostgreSQL connection pool stats and Redis cache status.
        Status is "healthy" when both DB and cache are reachable, "degraded" when cache is unreachable,
        and "unhealthy" when the database is unreachable.
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy or degraded (cache issue)
          schema:
            $ref: '#/definitions/internal_handler.HealthResponse'
        "503":
          description: Service is unhealthy (DB unreachable)
          schema:
            $ref: '#/definitions/internal_handler.HealthResponse'
      summary: Health check
      tags:
      - system
  /v1/signals:
    get:
      description: |-
        Returns a paginated list of trading signals. Scenarios and results are NOT included.
        Results are cached in Redis (TTL configurable, default 30s). Cache state is exposed via X-Cache header.
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Filter by trading pair, case-insensitive (e.g. btcusdt or BTCUSDT)
        in: query
        name: symbol
        type: string
      - description: Filter by source type (e.g. telegram, rss, http)
        in: query
        name: source_type
        type: string
      - description: Filter by source provider name
        in: query
        name: source_provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated signal list
          headers:
            X-Cache:
              description: 'Cache state: HIT or MISS'
              type: string
          schema:
            $ref: '#/definitions/signal_api_internal_repository.SignalsPage'
        "408":
          description: DB query exceeded 5s timeout
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: Global rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "503":
          description: Server at capacity – concurrency limit hit
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List signals
      tags:
      - signals
  /v1/signals/{symbol}:
    get:
      description: |-
        Returns a paginated list of signals filtered by the given trading pair symbol.
        Symbol matching is case-insensitive on input; stored values are uppercase (e.g. BTCUSDT).
      parameters:
      - description: Trading pair symbol (e.g. BTCUSDT)
        in: path
        name: symbol
        required: true
        type: string
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Filter by source type
        in: query
        name: source_type
        type: string
      - description: Filter by source provider
        in: query
        name: source_provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated signal list
          headers:
            X-Cache:
              description: 'Cache state: HIT or MISS'
              type: string
          schema:
            $ref: '#/definitions/signal_api_internal_repository.SignalsPage'
        "408":
          description: DB query exceeded 5s timeout
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: Global rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "503":
          description: Server at capacity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List signals by symbol
      tags:
      - signals
  /v1/signals/{symbol}/results:
    get:
      description: Returns a paginated list of signals for a specific trading pair,
        with full nested scenarios and results.
      parameters:
      - description: Trading pair symbol (e.g. BTCUSDT)
        in: path
        name: symbol
        required: true
        type: string
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Filter by source type
        in: query
        name: source_type
        type: string
      - description: Filter by source provider
        in: query
        name: source_provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated signal list with nested scenarios and results
          headers:
            X-Cache:
              description: 'Cache state: HIT or MISS'
              type: string
          schema:
            $ref: '#/definitions/signal_api_internal_repository.SignalsPage'
        "408":
          description: DB query exceeded 5s timeout
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: Global rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "503":
          description: Server at capacity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List signals by symbol with scenarios and results
      tags:
      - signals
  /v1/signals/results:
    get:
      description: |-
        Returns a paginated list of signals with full nested data: each Signal includes its Scenarios,
        and each Scenario includes its ScenarioResults. Uses 3 DB queries total (no N+1).
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Filter by trading pair
        in: query
        name: symbol
        type: string
      - description: Filter by source type
        in: query
        name: source_type
        type: string
      - description: Filter by source provider
        in: query
        name: source_provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated signal list with nested scenarios and results
          headers:
            X-Cache:
              description: 'Cache state: HIT or MISS'
              type: string
          schema:
            $ref: '#/definitions/signal_api_internal_repository.SignalsPage'
        "408":
          description: DB query exceeded 5s timeout
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: Global rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "503":
          description: Server at capacity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List signals with scenarios and results
      tags:
      - signals
  /v1/signals/source/{provider}:
    get:
      description: Returns a paginated list of signals filtered by the source provider
        name.
      parameters:
      - description: Source provider name (e.g. binance, coinbase, my_channel)
        in: path
        name: provider
        required: true
        type: string
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Filter by trading pair
        in: query
        name: symbol
        type: string
      - description: Filter by source type
        in: query
        name: source_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated signal list
          headers:
            X-Cache:
              description: 'Cache state: HIT or MISS'
              type: string
          schema:
            $ref: '#/definitions/signal_api_internal_repository.SignalsPage'
        "408":
          description: DB query exceeded 5s timeout
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: Global rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "503":
          description: Server at capacity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List signals by source provider
      tags:
      - signals
  /v1/signals/source/{provider}/results:
    get:
      description: Returns a paginated list of signals for a specific source provider,
        with full nested scenarios and results.
      parameters:
      - description: Source provider name
        in: path
        name: provider
        required: true
        type: string
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: Filter by trading pair
        in: query
        name: symbol
        type: string
      - description: Filter by source type
        in: query
        name: source_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated signal list with nested scenarios and results
          headers:
            X-Cache:
              description: 'Cache state: HIT or MISS'
              type: string
          schema:
            $ref: '#/definitions/signal_api_internal_repository.SignalsPage'
        "408":
          description: DB query exceeded 5s timeout
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: Global rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "503":
          description: Server at capacity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List signals by provider with scenarios and results
      tags:
      - signals
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Paginated trading signal endpoints with optional filtering by symbol,
    source type, and provider
  name: signals
- description: Service health and diagnostics
  name: system